requirements.txt
python-dotenv
langchain==0.2.1
langchain-community==0.2.1
langchain_chroma==0.1.4
langchain-openai
openai
requests
chromadb==0.5.0
pypdf==4.2.0
pydantic==2.10.3
pydantic-ai-slim[openai]==0.0.13